+++ actix-web/src/middleware/compress.rs
@@ -191,13 +191,10 @@ where
                             None => true,
                             Some(hdr) => {
                                 match hdr.to_str().ok().and_then(|hdr| hdr.parse::<Mime>().ok()) {
-                                    Some(mime)
-                                        if mime.type_().as_str() == "image"
-                                            && mime.subtype().as_str() != "svg+xml" =>
-                                    {
-                                        false
+                                    Some(mime) if mime.type_() == mime::IMAGE => {
+                                        matches!(mime.subtype(), mime::SVG)
                                     }
-                                    Some(mime) if mime.type_().as_str() == "video" => false,
+                                    Some(mime) if mime.type_() == mime::VIDEO => false,
                                     _ => true,
                                 }
                             }