+++ actix-web/CHANGES.md
@@ -6,6 +6,7 @@
 - Improve handling of non-UTF-8 header values in `Logger` middleware.
 - Add `HttpServer::shutdown_signal()` method.
 - Mark `HttpServer` as `#[must_use]`.
+- Allow SVG images to be compressed by the `Compress` middleware.
 - Re-export `mime` dependency.
 - Update `brotli` dependency to `8`.
 
@@ -28,7 +29,6 @@
 - On Windows, an error is now returned from `HttpServer::bind()` (or TLS variants) when binding to a socket that's already in use.
 - Update `brotli` dependency to `7`.
 - Minimum supported Rust version (MSRV) is now 1.75.
-- Fix Svg images not being compressed by the `Compress` middleware.
 
 ## 4.9.0